+++ io7m-jtensors-core/src/main/java/com/io7m/jtensors/MatrixI3x3D.java
@@ -48,9 +48,9 @@ import com.io7m.jnull.Nullable;
     for (int row = 0; row < 3; ++row) {
       for (int col = 0; col < 3; ++col) {
         if (row == col) {
-          m[row][col] = 1.0f;
+          m[row][col] = (double) 1.0;
         } else {
-          m[row][col] = 0.0f;
+          m[row][col] = (double) 0.0;
         }
       }
     }
