+++ io7m-jtensors-core/src/main/java/com/io7m/jtensors/MatrixM4x4D.java
@@ -442,7 +442,7 @@ public final class MatrixM4x4D implements
     final double r2c3 = m.getRowColumnD(2, 3);
     final double r3c3 = m.getRowColumnD(3, 3);
 
-    double sum = 0;
+    double sum = 0.0;
 
     sum += r0c0 * r1c1 * r2c2 * r3c3;
     sum -= r0c0 * r1c1 * r2c3 * r3c2;
@@ -698,7 +698,7 @@ public final class MatrixM4x4D implements
       return Option.none();
     }
 
-    final double d_inv = 1 / d;
+    final double d_inv = 1.0 / d;
 
     /**
      * This code is based on the Laplace Expansion theorem. Essentially, the
@@ -1263,7 +1263,7 @@ public final class MatrixM4x4D implements
 
     final double s = Math.sin(angle);
     final double c = Math.cos(angle);
-    final double t = 1 - c;
+    final double t = 1.0 - c;
 
     final double tx_sq = t * (axis_x * axis_x);
     final double ty_sq = t * (axis_y * axis_y);
@@ -1280,22 +1280,22 @@ public final class MatrixM4x4D implements
     final double r0c0 = tx_sq + c;
     final double r0c1 = txy - sz;
     final double r0c2 = txz + sy;
-    final double r0c3 = 0;
+    final double r0c3 = 0.0;
 
     final double r1c0 = txy + sz;
     final double r1c1 = ty_sq + c;
     final double r1c2 = tyz - sx;
-    final double r1c3 = 0;
+    final double r1c3 = 0.0;
 
     final double r2c0 = txz - sy;
     final double r2c1 = tyz + sx;
     final double r2c2 = tz_sq + c;
-    final double r2c3 = 0;
+    final double r2c3 = 0.0;
 
-    final double r3c0 = 0;
-    final double r3c1 = 0;
-    final double r3c2 = 0;
-    final double r3c3 = 1;
+    final double r3c0 = 0.0;
+    final double r3c1 = 0.0;
+    final double r3c2 = 0.0;
+    final double r3c3 = 1.0;
 
     out.setUnsafe(0, 0, r0c0);
     out.setUnsafe(0, 1, r0c1);
@@ -1409,12 +1409,12 @@ public final class MatrixM4x4D implements
     out.setUnsafe(0, 0, 1.0);
     out.setUnsafe(0, 1, 0.0);
     out.setUnsafe(0, 2, 0.0);
-    out.setUnsafe(0, 3, v.getXI());
+    out.setUnsafe(0, 3, (double) v.getXI());
 
     out.setUnsafe(1, 0, 0.0);
     out.setUnsafe(1, 1, 1.0);
     out.setUnsafe(1, 2, 0.0);
-    out.setUnsafe(1, 3, v.getYI());
+    out.setUnsafe(1, 3, (double) v.getYI());
 
     out.setUnsafe(2, 0, 0.0);
     out.setUnsafe(2, 1, 0.0);
@@ -1521,17 +1521,17 @@ public final class MatrixM4x4D implements
     out.setUnsafe(0, 0, 1.0);
     out.setUnsafe(0, 1, 0.0);
     out.setUnsafe(0, 2, 0.0);
-    out.setUnsafe(0, 3, v.getXI());
+    out.setUnsafe(0, 3, (double) v.getXI());
 
     out.setUnsafe(1, 0, 0.0);
     out.setUnsafe(1, 1, 1.0);
     out.setUnsafe(1, 2, 0.0);
-    out.setUnsafe(1, 3, v.getYI());
+    out.setUnsafe(1, 3, (double) v.getYI());
 
     out.setUnsafe(2, 0, 0.0);
     out.setUnsafe(2, 1, 0.0);
     out.setUnsafe(2, 2, 1.0);
-    out.setUnsafe(2, 3, v.getZI());
+    out.setUnsafe(2, 3, (double) v.getZI());
 
     out.setUnsafe(3, 0, 0.0);
     out.setUnsafe(3, 1, 0.0);
@@ -1559,97 +1559,97 @@ public final class MatrixM4x4D implements
     final MatrixReadable4x4DType m1,
     final MatrixM4x4D out)
   {
-    double r0c0 = 0;
+    double r0c0 = 0.0;
     r0c0 += m0.getRowColumnD(0, 0) * m1.getRowColumnD(0, 0);
     r0c0 += m0.getRowColumnD(0, 1) * m1.getRowColumnD(1, 0);
     r0c0 += m0.getRowColumnD(0, 2) * m1.getRowColumnD(2, 0);
     r0c0 += m0.getRowColumnD(0, 3) * m1.getRowColumnD(3, 0);
 
-    double r1c0 = 0;
+    double r1c0 = 0.0;
     r1c0 += m0.getRowColumnD(1, 0) * m1.getRowColumnD(0, 0);
     r1c0 += m0.getRowColumnD(1, 1) * m1.getRowColumnD(1, 0);
     r1c0 += m0.getRowColumnD(1, 2) * m1.getRowColumnD(2, 0);
     r1c0 += m0.getRowColumnD(1, 3) * m1.getRowColumnD(3, 0);
 
-    double r2c0 = 0;
+    double r2c0 = 0.0;
     r2c0 += m0.getRowColumnD(2, 0) * m1.getRowColumnD(0, 0);
     r2c0 += m0.getRowColumnD(2, 1) * m1.getRowColumnD(1, 0);
     r2c0 += m0.getRowColumnD(2, 2) * m1.getRowColumnD(2, 0);
     r2c0 += m0.getRowColumnD(2, 3) * m1.getRowColumnD(3, 0);
 
-    double r3c0 = 0;
+    double r3c0 = 0.0;
     r3c0 += m0.getRowColumnD(3, 0) * m1.getRowColumnD(0, 0);
     r3c0 += m0.getRowColumnD(3, 1) * m1.getRowColumnD(1, 0);
     r3c0 += m0.getRowColumnD(3, 2) * m1.getRowColumnD(2, 0);
     r3c0 += m0.getRowColumnD(3, 3) * m1.getRowColumnD(3, 0);
 
-    double r0c1 = 0;
+    double r0c1 = 0.0;
     r0c1 += m0.getRowColumnD(0, 0) * m1.getRowColumnD(0, 1);
     r0c1 += m0.getRowColumnD(0, 1) * m1.getRowColumnD(1, 1);
     r0c1 += m0.getRowColumnD(0, 2) * m1.getRowColumnD(2, 1);
     r0c1 += m0.getRowColumnD(0, 3) * m1.getRowColumnD(3, 1);
 
-    double r1c1 = 0;
+    double r1c1 = 0.0;
     r1c1 += m0.getRowColumnD(1, 0) * m1.getRowColumnD(0, 1);
     r1c1 += m0.getRowColumnD(1, 1) * m1.getRowColumnD(1, 1);
     r1c1 += m0.getRowColumnD(1, 2) * m1.getRowColumnD(2, 1);
     r1c1 += m0.getRowColumnD(1, 3) * m1.getRowColumnD(3, 1);
 
-    double r2c1 = 0;
+    double r2c1 = 0.0;
     r2c1 += m0.getRowColumnD(2, 0) * m1.getRowColumnD(0, 1);
     r2c1 += m0.getRowColumnD(2, 1) * m1.getRowColumnD(1, 1);
     r2c1 += m0.getRowColumnD(2, 2) * m1.getRowColumnD(2, 1);
     r2c1 += m0.getRowColumnD(2, 3) * m1.getRowColumnD(3, 1);
 
-    double r3c1 = 0;
+    double r3c1 = 0.0;
     r3c1 += m0.getRowColumnD(3, 0) * m1.getRowColumnD(0, 1);
     r3c1 += m0.getRowColumnD(3, 1) * m1.getRowColumnD(1, 1);
     r3c1 += m0.getRowColumnD(3, 2) * m1.getRowColumnD(2, 1);
     r3c1 += m0.getRowColumnD(3, 3) * m1.getRowColumnD(3, 1);
 
-    double r0c2 = 0;
+    double r0c2 = 0.0;
     r0c2 += m0.getRowColumnD(0, 0) * m1.getRowColumnD(0, 2);
     r0c2 += m0.getRowColumnD(0, 1) * m1.getRowColumnD(1, 2);
     r0c2 += m0.getRowColumnD(0, 2) * m1.getRowColumnD(2, 2);
     r0c2 += m0.getRowColumnD(0, 3) * m1.getRowColumnD(3, 2);
 
-    double r1c2 = 0;
+    double r1c2 = 0.0;
     r1c2 += m0.getRowColumnD(1, 0) * m1.getRowColumnD(0, 2);
     r1c2 += m0.getRowColumnD(1, 1) * m1.getRowColumnD(1, 2);
     r1c2 += m0.getRowColumnD(1, 2) * m1.getRowColumnD(2, 2);
     r1c2 += m0.getRowColumnD(1, 3) * m1.getRowColumnD(3, 2);
 
-    double r2c2 = 0;
+    double r2c2 = 0.0;
     r2c2 += m0.getRowColumnD(2, 0) * m1.getRowColumnD(0, 2);
     r2c2 += m0.getRowColumnD(2, 1) * m1.getRowColumnD(1, 2);
     r2c2 += m0.getRowColumnD(2, 2) * m1.getRowColumnD(2, 2);
     r2c2 += m0.getRowColumnD(2, 3) * m1.getRowColumnD(3, 2);
 
-    double r3c2 = 0;
+    double r3c2 = 0.0;
     r3c2 += m0.getRowColumnD(3, 0) * m1.getRowColumnD(0, 2);
     r3c2 += m0.getRowColumnD(3, 1) * m1.getRowColumnD(1, 2);
     r3c2 += m0.getRowColumnD(3, 2) * m1.getRowColumnD(2, 2);
     r3c2 += m0.getRowColumnD(3, 3) * m1.getRowColumnD(3, 2);
 
-    double r0c3 = 0;
+    double r0c3 = 0.0;
     r0c3 += m0.getRowColumnD(0, 0) * m1.getRowColumnD(0, 3);
     r0c3 += m0.getRowColumnD(0, 1) * m1.getRowColumnD(1, 3);
     r0c3 += m0.getRowColumnD(0, 2) * m1.getRowColumnD(2, 3);
     r0c3 += m0.getRowColumnD(0, 3) * m1.getRowColumnD(3, 3);
 
-    double r1c3 = 0;
+    double r1c3 = 0.0;
     r1c3 += m0.getRowColumnD(1, 0) * m1.getRowColumnD(0, 3);
     r1c3 += m0.getRowColumnD(1, 1) * m1.getRowColumnD(1, 3);
     r1c3 += m0.getRowColumnD(1, 2) * m1.getRowColumnD(2, 3);
     r1c3 += m0.getRowColumnD(1, 3) * m1.getRowColumnD(3, 3);
 
-    double r2c3 = 0;
+    double r2c3 = 0.0;
     r2c3 += m0.getRowColumnD(2, 0) * m1.getRowColumnD(0, 3);
     r2c3 += m0.getRowColumnD(2, 1) * m1.getRowColumnD(1, 3);
     r2c3 += m0.getRowColumnD(2, 2) * m1.getRowColumnD(2, 3);
     r2c3 += m0.getRowColumnD(2, 3) * m1.getRowColumnD(3, 3);
 
-    double r3c3 = 0;
+    double r3c3 = 0.0;
     r3c3 += m0.getRowColumnD(3, 0) * m1.getRowColumnD(0, 3);
     r3c3 += m0.getRowColumnD(3, 1) * m1.getRowColumnD(1, 3);
     r3c3 += m0.getRowColumnD(3, 2) * m1.getRowColumnD(2, 3);
@@ -2111,9 +2111,9 @@ public final class MatrixM4x4D implements
     for (int row = 0; row < MatrixM4x4D.VIEW_ROWS; ++row) {
       for (int col = 0; col < MatrixM4x4D.VIEW_COLS; ++col) {
         if (row == col) {
-          m.setUnsafe(row, col, 1.0f);
+          m.setUnsafe(row, col, (double) 1.0);
         } else {
-          m.setUnsafe(row, col, 0.0f);
+          m.setUnsafe(row, col, (double) 0.0);
         }
       }
     }
